+++ docs/ProcessManager.Client/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,9 @@
 # ProcessManager.Client Release Notes
 
+## Version 0.27.1
+
+- No functional changes.
+
 ## Version 0.27.0
 
 - Update Nuget package `Grpc.Tools`