+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,10 @@
 # ProcessManager.Orchestrations.Abstractions Release Notes
 
+## Version 0.20.0
+
+- Moved all types from namespace `ValueObjects` to new NuGet package `Energinet.DataHub.ProcessManager.Components.ValueObjects`
+- Add reference to new NuGet package `Energinet.DataHub.ProcessManager.Components.ValueObjects`
+
 ## Version 0.19.0
 
 - Add protobuf contract `CalculationEnqueueCompletedV1`